+++ src/Traits/CanBePaid.php
@@ -72,17 +72,25 @@ trait CanBePaid
 
     /**
      * @param Product $product
+     * @param bool $gifts
      * @return null|Transfer
      */
-    public function paid(Product $product): ?Transfer
+    public function paid(Product $product, bool $gifts = false): ?Transfer
     {
+        $status = [Transfer::STATUS_PAID];
+        if ($gifts) {
+            $status[] = Transfer::STATUS_GIFT;
+        }
+        
         /**
          * @var Model $product
+         * @var Transfer $query
          */
-        return $this->transfers()
+        $query = $this->transfers();
+        return $query
             ->where('to_type', $product->getMorphClass())
             ->where('to_id', $product->getKey())
-            ->where('refund', 0)
+            ->whereIn('status', $status)
             ->orderBy('id', 'desc')
             ->first();
     }
@@ -90,12 +98,13 @@ trait CanBePaid
     /**
      * @param Product $product
      * @param bool $force
+     * @param bool $gifts
      * @return bool
      * @throws
      */
-    public function refund(Product $product, bool $force = false): bool
+    public function refund(Product $product, bool $force = false, bool $gifts = false): bool
     {
-        $transfer = $this->paid($product);
+        $transfer = $this->paid($product, $gifts);
 
         if (!$transfer) {
             throw (new ModelNotFoundException())
@@ -103,25 +112,39 @@ trait CanBePaid
         }
 
         return DB::transaction(function () use ($product, $transfer, $force) {
+            $transfer->load('withdraw.payable');
+
             if ($force) {
-                $product->forceTransfer($this, $transfer->deposit->amount, $product->getMetaProduct());
+                $product->forceTransfer(
+                    $transfer->withdraw->payable,
+                    $transfer->deposit->amount,
+                    $product->getMetaProduct()
+                );
             } else {
-                $product->transfer($this, $transfer->deposit->amount, $product->getMetaProduct());
+                $product->transfer(
+                    $transfer->withdraw->payable,
+                    $transfer->deposit->amount,
+                    $product->getMetaProduct()
+                );
             }
 
-            return $transfer->update(['refund' => 1]);
+            return $transfer->update([
+                'status' => Transfer::STATUS_REFUND,
+                'status_last' => $transfer->status,
+            ]);
         });
     }
 
     /**
      * @param Product $product
      * @param bool $force
+     * @param bool $gifts
      * @return bool
      */
-    public function safeRefund(Product $product, bool $force = false): bool
+    public function safeRefund(Product $product, bool $force = false, bool $gifts = false): bool
     {
         try {
-            return $this->refund($product, $force);
+            return $this->refund($product, $force, $gifts);
         } catch (\Throwable $throwable) {
             return false;
         }
@@ -129,12 +152,13 @@ trait CanBePaid
 
     /**
      * @param Product $product
+     * @param bool $gifts
      * @return bool
      * @throws
      */
-    public function forceRefund(Product $product): bool
+    public function forceRefund(Product $product, bool $gifts = false): bool
     {
-        return $this->refund($product, true);
+        return $this->refund($product, true, $gifts);
     }
 
 }